# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mass Polimer ERP API",
    "version": "1.0.0",
...
    "license": {
      "name": "UNLICENSED"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
  "security": [
    {
      "FirebaseAuth": []
    },
    {
      "DevUser": []
    }
  ],
  "paths": {
...
    "/api/me/permissions": {
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "getMyPermissions",
        "summary": "Get the caller’s effective access",
        "description": "Role screens ± per-employee grants. The client gates its menu on this; the server enforces the same policy independently.",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The caller’s admin flag and screen list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isAdmin": {
                      "type": "boolean"
                    },
                    "screens": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/directory": {
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "listDirectory",
        "summary": "List the member directory",
        "description": "Backs the login picker and role switcher, so it is readable by any authenticated member.",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The member roster.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      },
                      "employeeCode": {
                        "type": "string"
                      },
                      "department": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/screens": {
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "listScreens",
        "summary": "List the screen catalog",
        "description": "Every feature key that can be granted, for the roles editor.\n\n**Requires permission:** `screen:acl`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The catalog of screen keys.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "screens": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:acl`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/employees": {
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "listEmployees",
        "summary": "List employees",
        "description": "**Requires permission:** `screen:users`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employees in the caller’s organization.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Membership"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "user": {
                            "$ref": "#/components/schemas/User",
                            "description": "A User; this endpoint may select only some of its fields."
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:users`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Administration"
        ],
        "operationId": "createEmployee",
        "summary": "Add an employee",
        "description": "**Requires permission:** `screen:users`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "roleId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "department": {
                    "type": "string"
                  },
                  "shift": {
                    "type": "string"
                  },
                  "employeeCode": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "on_leave",
                      "inactive"
                    ]
                  }
                },
                "required": [
                  "name",
                  "email",
                  "roleId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created employee membership.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Membership"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User",
                          "description": "A User; this endpoint may select only some of its fields."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:users`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "- `already_member` — That person is already an employee of this organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "- `bad_role` — The referenced role does not exist in this organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/employees/{id}": {
      "patch": {
        "tags": [
          "Administration"
        ],
        "operationId": "updateEmployee",
        "summary": "Update an employee",
        "description": "**Requires permission:** `screen:users`",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Membership id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "roleId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "department": {
                    "type": "string"
                  },
                  "shift": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "on_leave",
                      "inactive"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated employee.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Membership"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/User",
                          "description": "A User; this endpoint may select only some of its fields."
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:users`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "- `not_found` — No employee with that id exists in your organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "- `bad_role` — The referenced role does not exist in this organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/roles": {
      "get": {
        "tags": [
          "Administration"
        ],
        "operationId": "listRoles",
        "summary": "List roles",
        "description": "**Requires permission:** `screen:users`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Built-in and custom roles, each with a count of the employees on it.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Role"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "_count": {
                            "type": "object",
                            "description": "Derived `_count` object attached by this endpoint."
                          }
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:users`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Administration"
        ],
        "operationId": "createRole",
        "summary": "Create a custom role",
        "description": "**Requires permission:** `screen:acl`",
        "parameters": [
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "screens": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "default": []
                  }
                },
                "required": [
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:acl`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "- `name_taken` — A role with that name already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "`validation` — the body failed schema validation. `error.details` carries the field errors.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/roles/{id}": {
      "patch": {
        "tags": [
          "Administration"
        ],
        "operationId": "updateRole",
        "summary": "Update a role",
        "description": "**Requires permission:** `screen:acl`",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "screens": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:acl`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "- `not_found` — No role with that id exists in your organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "- `system_role` — Built-in roles cannot be renamed.\n- `name_taken` — A role with that name already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "`validation` — the body failed schema validation. `error.details` carries the field errors.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Administration"
        ],
        "operationId": "deleteRole",
        "summary": "Delete a custom role",
        "description": "**Requires permission:** `screen:acl`",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Role id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dev-org",
            "in": "header",
            "required": false,
            "description": "Development only: pick the organization by id or slug when the identity has more than one membership.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmation that the role was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthenticated` — no usable credential, or no membership resolved for it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "`forbidden` — the caller's role does not hold `screen:acl`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "- `not_found` — No role with that id exists in your organization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "- `system_role` — Built-in roles cannot be deleted.\n- `role_in_use` — Employees are still assigned to the role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — unhandled server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "FirebaseAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token, verified server-side. This is the production scheme; it activates when `DEV_AUTH=0` and the Firebase Admin credentials are configured."
      },
      "DevUser": {
        "type": "apiKey",
        "in": "header",
        "name": "x-dev-user",
        "description": "Development only, active while `DEV_AUTH` is on. Pass an employee code or email to act as that member; omit it to fall back to a seeded Administrator. Pair with `x-dev-org` to choose the organization. Never enable this in production."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The envelope every failure is returned in.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine-readable code. Branch on this, not on the message."
              },
              "message": {
                "type": "string",
                "description": "Human-readable explanation, safe to surface in a UI."
              },
              "details": {
                "description": "Optional context. For `validation` this is the Zod field-error map."
              }
            }
          }
        }
      },
...
      "Membership": {
        "type": "object",
        "description": "The Membership record as stored. Relation objects appear only on endpoints that include them.",
        "properties": {
          "id": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "employeeCode": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "shift": {
            "type": "string"
          },
          "line": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "joinDate": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "lastSeen": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "roleId": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "organizationId",
          "userId",
          "employeeCode",
          "department",
          "role",
          "shift",
          "line",
          "status",
          "joinDate",
          "location",
          "lastSeen",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}